Refactor action imports to align with new directory structure in list module
