+++ src/app/list/actions/remove-torrent.ts
